refactor: move task components
