+++ src/app/app.module.ts
@@ -3,7 +3,7 @@ import {BrowserModule} from '@angular/platform-browser';
 
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
-import {HeaderComponent} from './components/header/header.component';
+import {HeaderComponent} from './shared/components/header/header.component';
 import {NgOptimizedImage} from "@angular/common";
 import {NgIconsModule} from "@ng-icons/core";
 import {
@@ -19,12 +19,9 @@ import {
     heroSun,
     heroXMark
 } from "@ng-icons/heroicons/outline";
-import {TimerComponent} from './components/timer/timer.component';
-import {TasksComponent} from './components/tasks/tasks.component';
-import {CreateTaskComponent} from './components/tasks/components/create-task/create-task.component';
+import {TimerComponent} from './pages/home/components/timer/timer.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import {TaskModalComponent} from './components/tasks/components/task-modal/task-modal.component';
-import {FooterComponent} from './components/footer/footer.component';
+import {FooterComponent} from './shared/components/footer/footer.component';
 import {ShowTimePipe} from './pipe/show-time.pipe';
 import {TasksService} from "./shared/services/tasks.service";
 import {SettingsComponent} from './pages/settings/settings.component';
@@ -38,6 +35,9 @@ import {
 import {ControlSettingsComponent} from './pages/settings/components/control-settings/control-settings.component';
 import {DarkmodeService} from "./shared/services/darkmode.service";
 import {BadgeComponent} from './shared/components/badge/badge.component';
+import {TasksComponent} from "./pages/home/components/tasks/tasks.component";
+import {CreateTaskComponent} from "./pages/home/components/tasks/components/create-task/create-task.component";
+import {TaskModalComponent} from "./pages/home/components/tasks/components/task-modal/task-modal.component";
 
 @NgModule({
     declarations: [
@@ -55,7 +55,7 @@ import {BadgeComponent} from './shared/components/badge/badge.component';
         ShowMinutesPipe,
         NotificationSettingsComponent,
         ControlSettingsComponent,
-        BadgeComponent
+        BadgeComponent,
     ],
     imports: [
         BrowserModule,
