--- src/app/components/tasks/components/create-task/create-task.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { CreateTaskComponent } from './create-task.component';
-
-describe('CreateTaskComponent', () => {
-  let component: CreateTaskComponent;
-  let fixture: ComponentFixture<CreateTaskComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [CreateTaskComponent]
-    });
-    fixture = TestBed.createComponent(CreateTaskComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
