--- src/app/components/tasks/tasks.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { TasksComponent } from './tasks.component';
-
-describe('TasksComponent', () => {
-  let component: TasksComponent;
-  let fixture: ComponentFixture<TasksComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [TasksComponent]
-    });
-    fixture = TestBed.createComponent(TasksComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
