+++ src/app/pages/home/components/tasks/components/task-modal/task-modal.component.spec.ts
@@ -1,0 +1,21 @@
+import {ComponentFixture, TestBed} from '@angular/core/testing';
+
+import {TaskModalComponent} from './task-modal.component';
+
+describe('TaskModalComponent', () => {
+    let component: TaskModalComponent;
+    let fixture: ComponentFixture<TaskModalComponent>;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            declarations: [TaskModalComponent]
+        });
+        fixture = TestBed.createComponent(TaskModalComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+});
