+++ src/app/pages/home/components/timer/timer.component.html
