+++ src/app/pages/home/components/timer/timer.component.spec.ts
@@ -1,0 +1,21 @@
+import {ComponentFixture, TestBed} from '@angular/core/testing';
+
+import {TimerComponent} from './timer.component';
+
+describe('ChronoComponent', () => {
+    let component: TimerComponent;
+    let fixture: ComponentFixture<TimerComponent>;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            declarations: [TimerComponent]
+        });
+        fixture = TestBed.createComponent(TimerComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+});
