+++ src/app/shared/components/footer/footer.component.ts
@@ -1,0 +1,10 @@
+import {Component} from '@angular/core';
+
+@Component({
+    selector: 'app-footer',
+    templateUrl: './footer.component.html',
+    styleUrls: ['./footer.component.scss']
+})
+export class FooterComponent {
+
+}
